+++ src/include/gameActor.h
@@ -6,31 +6,31 @@
 class GameActor {
     public:
         GameActor();
-        GameActor(int x, int y, int vx, int vy, int speed, Controller* c);
+        GameActor(double x, double y, double vx, double vy, double speed, Controller* c);
         ~GameActor();
         virtual void UpdatePosition();
         virtual void Draw();
         virtual void Update();
 
         // Getters and setters
-        inline void setX(int x) { x_ = x; }
-        inline void setY(int y) { y_ = y; }
-        inline void setVelX(int v) { vel_x_ = v; }
-        inline void setVelY(int v) { vel_y_ = v; }
+        inline void setX(double x) { x_ = x; }
+        inline void setY(double y) { y_ = y; }
+        inline void setVelX(double v) { vel_x_ = v; }
+        inline void setVelY(double v) { vel_y_ = v; }
         inline void setSpeed(int speed) { speed_ = speed; }
-        inline int getX() { return x_; }
-        inline int getY() { return y_; }
-        inline int getVelX() { return vel_x_; }
-        inline int getVelY() { return vel_y_; }
-        inline int getSpeed() { return speed_; }
+        inline double getX() { return x_; }
+        inline double getY() { return y_; }
+        inline double getVelX() { return vel_x_; }
+        inline double getVelY() { return vel_y_; }
+        inline double getSpeed() { return speed_; }
         inline Controller* getController() { return controller_; }
 
     protected:
-        int x_;
-        int y_;
-        int vel_x_;
-        int vel_y_;
-        int speed_;
+        double x_;
+        double y_;
+        double vel_x_;
+        double vel_y_;
+        double speed_;
         Controller* controller_;
 };
